refactor(auth): 方案 B — request() 移入 service 层 + ScuAuthService 拆分
- ScuAuthService 拆分到 lib/services/scu_auth/ 文件夹
  - scu_auth_service.dart — 主类（认证核心 + fetchXxx + request()）
  - scu_auth_models.dart — CaptchaResult + ScuLoginException + kZhjwBase
  - cookie_client.dart — CookieClient 类
  - 旧路径 scu_auth_service.dart 改为 barrel export
- fetchXxx 方法内部使用 _authManager.scu.request()，移除 {CookieClient? client} 参数
- 新增 service.request() 通用包装供非 fetchXxx 调用方使用
- AuthManager 构造函数调 scu.service.bindAuthManager(this)
- GradesProvider / PlanCompletionProvider / TrainProgramProvider 移除 AuthManager 依赖
- ClassroomPage / CourseScheduleSetting / ImportSchedulePage 直接调 service.fetchXxx()
- ProfileLabelsProvider 改用 service.request()
- injector.dart 简化注入逻辑
