+++ lib/services/scu_auth_service.dart
@@ -1,721 +1,4 @@
-import 'dart:convert';
-import 'package:http/http.dart' as http;
-import 'package:bugaoshan/pages/campus/models/classroom_model.dart';
-import 'package:bugaoshan/utils/constants.dart';
-import 'package:bugaoshan/utils/json_utils.dart';
-import 'package:bugaoshan/utils/sm2_crypto.dart';
-
-/// 教务系统 base URL（该服务器不支持 HTTPS）
-const kZhjwBase = 'http://zhjw.scu.edu.cn';
-
-/// 四川大学统一身份认证 Service
-class ScuAuthService {
-  static const _base = 'https://id.scu.edu.cn';
-  static const _clientId = '1371cbeda563697537f28d99b4744a973uDKtgYqL5B';
-  static const _enterpriseId = 'scdx';
-
-  static final Map<String, String> _headers = {
-    'Accept': 'application/json, text/plain, */*',
-    'Content-Type': 'application/json;charset=UTF-8',
-    'Origin': _base,
-    'Referer': '$_base/frontend/login',
-    'User-Agent': kDefaultUserAgent,
-  };
-
-  String? _accessToken;
-  String? get accessToken => _accessToken;
-
-  CookieClient? _cachedClient;
-  Future<CookieClient>? _bindSessionFuture;
-
-  void restoreAccessToken(String? token) {
-    _accessToken = token;
-    _cachedClient = null;
-  }
-
-  /// 清除缓存的 Client，下次 [bindSession] 会重新执行 SSO 握手。
-  void invalidateCachedClient() {
-    _cachedClient = null;
-  }
-
-  /// 获取验证码，返回 [CaptchaResult]
-  Future<CaptchaResult> fetchCaptcha() async {
-    final ts = DateTime.now().millisecondsSinceEpoch;
-    final uri = Uri.parse(
-      '$_base/api/public/bff/v1.2/one_time_login/captcha'
-      '?_enterprise_id=$_enterpriseId&timestamp=$ts',
-    );
-    final resp = await http.get(uri, headers: _headers).timeout(kHttpTimeout);
-
-    final json = parseJson(
-      resp.body,
-      'captcha',
-      (msg) => ScuLoginException(msg),
-    );
-    final data = json['data'];
-    if (data == null) {
-      throw ScuLoginException('验证码接口返回异常: ${resp.body}');
-    }
-    final dataMap = data as Map<String, dynamic>;
-
-    final captchaImg =
-        (dataMap['captcha'] ??
-                dataMap['image'] ??
-                dataMap['img'] ??
-                dataMap['captchaImage'])
-            ?.toString();
-    final code = dataMap['code']?.toString();
-
-    if (captchaImg == null || code == null) {
-      throw ScuLoginException('验证码字段解析失败，实际响应: ${resp.body}');
-    }
-    return CaptchaResult(code: code, captchaBase64: captchaImg);
-  }
-
-  /// 登录，成功后 [accessToken] 会被保存
-  Future<void> login({
-    required String username,
-    required String password,
-    required String captchaCode,
-    required String captchaText,
-  }) async {
-    // 1. 获取 SM2 公钥（服务端偶发 500，加重试）
-    Map<String, dynamic>? sm2Data;
-    String? lastSm2Body;
-    for (int attempt = 0; attempt < 3; attempt++) {
-      final sm2Resp = await http
-          .post(
-            Uri.parse('$_base/api/public/bff/v1.2/sm2_key'),
-            headers: _headers,
-            body: '{}',
-          )
-          .timeout(kHttpTimeout);
-      lastSm2Body = sm2Resp.body;
-      final sm2Json = parseJson(
-        sm2Resp.body,
-        'sm2_key',
-        (msg) => ScuLoginException(msg),
-      );
-      sm2Data = sm2Json['data'] as Map<String, dynamic>?;
-      if (sm2Data != null &&
-          sm2Data['publicKey'] != null &&
-          sm2Data['code'] != null) {
-        break;
-      }
-      if (attempt < 2) {
-        await Future.delayed(const Duration(milliseconds: 500));
-      }
-    }
-    if (sm2Data == null) {
-      throw ScuLoginException('SM2 公钥接口返回异常: $lastSm2Body');
-    }
-    final publicKey = sm2Data['publicKey']?.toString();
-    final sm2Code = sm2Data['code']?.toString();
-    if (publicKey == null || sm2Code == null) {
-      throw ScuLoginException('SM2 公钥字段缺失: $lastSm2Body');
-    }
-
-    // 2. SM2 C1C2C3 加密密码
-    final encryptedPassword = SM2Crypto.encryptWithBase64Key(
-      password,
-      publicKey,
-    );
-
-    // 3. 请求 token
-    final payload = jsonEncode({
-      'client_id': _clientId,
-      'grant_type': 'password',
-      'scope': 'read',
-      'username': username,
-      'password': encryptedPassword,
-      '_enterprise_id': _enterpriseId,
-      'sm2_code': sm2Code,
-      'cap_code': captchaCode,
-      'cap_text': captchaText,
-    });
-
-    final tokenResp = await http
-        .post(
-          Uri.parse('$_base/api/public/bff/v1.2/rest_token'),
-          headers: _headers,
-          body: payload,
-        )
-        .timeout(kHttpTimeout);
-
-    final result = parseJson(
-      tokenResp.body,
-      'rest_token',
-      (msg) => ScuLoginException(msg),
-    );
-    if (result['success'] != true) {
-      final msg =
-          result['message']?.toString() ?? result['msg']?.toString() ?? '登录失败';
-      throw ScuLoginException(msg);
-    }
-
-    final tokenData = result['data'] as Map<String, dynamic>?;
-    final token = tokenData?['access_token']?.toString();
-    if (token == null) {
-      throw ScuLoginException('Token 字段缺失: ${tokenResp.body}');
-    }
-
-    // 登录成功，清掉旧 client，新 token 需要重新建立 session
-    _accessToken = token;
-    _cachedClient = null;
-  }
-
-  /// 将 token 绑定到服务端 session，返回携带 cookie 的 Client。
-  ///
-  /// 预热两个 SSO SP：
-  ///   - scdxplugin_jwt23       → 教务系统（zhjw.scu.edu.cn）
-  ///   - scdxplugin_cas_apereo17 → 团委系统（CCYL / dekt.scu.edu.cn）
-  Future<CookieClient> bindSession() async {
-    if (_accessToken == null) throw ScuLoginException('未登录');
-
-    if (_cachedClient != null) return _cachedClient!;
-
-    // 并发保护：多个调用者同时 bindSession 时，只执行一次 SSO 握手
-    if (_bindSessionFuture != null) return _bindSessionFuture!;
-
-    _bindSessionFuture = _doBindSession();
-    try {
-      return await _bindSessionFuture!;
-    } finally {
-      _bindSessionFuture = null;
-    }
-  }
-
-  Future<CookieClient> _doBindSession() async {
-    final client = CookieClient();
-
-    // ── Step 1: 保存 token 到服务端 session（必须成功）──────────────────────
-    final sessionResp = await client.post(
-      Uri.parse('$_base/api/bff/v1.2/commons/session/save'),
-      headers: {..._headers, 'Authorization': 'Bearer $_accessToken'},
-      body: '{}',
-    );
-    final sessionResult = parseJson(
-      sessionResp.body,
-      'session/save',
-      (msg) => ScuLoginException(msg),
-    );
-    if (sessionResult['success'] != true) {
-      throw ScuLoginException('session/save 失败: ${sessionResp.body}');
-    }
-
-    // ── Step 2 & 3: 预热 JWT SSO + CAS Apereo SSO（相互独立，并行执行）─────
-    // JWT SSO 为教务系统（zhjw.scu.edu.cn）准备，夜间非校园网可能超时。
-    // CAS Apereo SSO 为 CCYL 团委系统准备。
-    // 两者任一失败均不阻断整体流程。
-    await Future.wait([_warmupJwtSso(client), _warmupCasSso(client)]);
-
-    client._reusable = true;
-    _cachedClient = client;
-    return client;
-  }
-
-  Future<void> _warmupJwtSso(CookieClient client) async {
-    try {
-      await client.followRedirects(
-        Uri.parse(
-          '$_base/enduser/sp/sso/scdxplugin_jwt23'
-          '?enterpriseId=scdx&target_url=index',
-        ),
-        headers: {
-          'Accept': 'text/html,application/xhtml+xml,*/*',
-          'User-Agent': _headers['User-Agent']!,
-          'Authorization': 'Bearer $_accessToken',
-        },
-      );
-    } catch (_) {
-      // JWT SSO 预热失败不影响非教务功能
-    }
-  }
-
-  Future<void> _warmupCasSso(CookieClient client) async {
-    try {
-      await client.followRedirects(
-        Uri.parse(
-          '$_base/api/bff/v1.2/commons/sp_logged'
-          '?access_token=$_accessToken'
-          '&sp_code=$kCcylSpCode'
-          '&application_key=scdxplugin_cas_apereo17',
-        ),
-        headers: {
-          'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*',
-          'User-Agent': _headers['User-Agent']!,
-        },
-      );
-    } catch (_) {
-      // CAS SSO 预热失败不影响非 CCYL 功能
-    }
-  }
-
-  /// 从教务系统首页获取当前教学周数
-  Future<int> fetchCurrentWeek({CookieClient? client}) async {
-    client ??= await bindSession();
-    final resp = await client.get(
-      Uri.parse('$kZhjwBase/'),
-      headers: {
-        'Accept': 'text/html,*/*',
-        'Referer': '$kZhjwBase/',
-        'User-Agent': _headers['User-Agent']!,
-      },
-    );
-    final body = resp.body.trim();
-    _checkSessionExpiry(body, resp.statusCode);
-    // 匹配 "2025-2026 春  第8周   星期五" 中的周数
-    final match = RegExp(r'第(\d+)周').firstMatch(body);
-    if (match == null) {
-      throw ScuLoginException('无法获取当前周数，请检查教务系统状态');
-    }
-    return int.parse(match.group(1)!);
-  }
-
-  /// 获取历年学期列表（需要先登录），返回 [{value: '2025-2026-2-1', label: '2025-2026学年春(当前)'}, ...]
-  Future<List<({String value, String label})>> fetchSemesters({
-    CookieClient? client,
-  }) async {
-    client ??= await bindSession();
-    final resp = await client.get(
-      Uri.parse(
-        '$kZhjwBase/student/courseSelect'
-        '/calendarSemesterCurriculum/index',
-      ),
-      headers: {
-        'Accept': 'text/html,*/*',
-        'Referer': '$kZhjwBase/',
-        'User-Agent': _headers['User-Agent']!,
-      },
-    );
-    final body = resp.body.trim();
-    _checkSessionExpiry(body, resp.statusCode);
-    final regex = RegExp(
-      r'<option[^>]+value="([^"]+)"[^>]*>(.*?)</option>',
-      dotAll: true,
-    );
-    final matches = regex.allMatches(body);
-    final semesters = matches.map((m) {
-      final value = m.group(1)!.trim();
-      final label = m.group(2)!.replaceAll(RegExp(r'<[^>]+>'), '').trim();
-      return (value: value, label: label);
-    }).toList();
-    if (semesters.isEmpty) {
-      throw ScuLoginException('无法获取学期列表，请检查登录状态');
-    }
-    return semesters;
-  }
-
-  /// 获取指定学期课表 JSON（需要先登录），[planCode] 如 '2025-2026-2-1'
-  Future<Map<String, dynamic>> fetchJwxtSchedule({
-    required String planCode,
-    CookieClient? client,
-  }) async {
-    client ??= await bindSession();
-    final resp = await client.post(
-      Uri.parse(
-        '$kZhjwBase/student/courseSelect'
-        '/thisSemesterCurriculum/ajaxStudentSchedule/callback',
-      ),
-      headers: {
-        'Accept': 'application/json, text/javascript, */*; q=0.01',
-        'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8',
-        'Referer':
-            '$kZhjwBase/student/courseSelect/calendarSemesterCurriculum/index',
-        'User-Agent': _headers['User-Agent']!,
-        'X-Requested-With': 'XMLHttpRequest',
-      },
-      body: 'planCode=$planCode',
-    );
-    final body = resp.body.trim();
-    _checkSessionExpiry(body, resp.statusCode);
-    return parseJson(body, 'jwxt/schedule', (msg) => ScuLoginException(msg));
-  }
-
-  /// 获取及格成绩（需要先登录）
-  Future<Map<String, dynamic>> fetchPassingScores({
-    CookieClient? client,
-  }) async {
-    client ??= await bindSession();
-    final indexResp = await client.get(
-      Uri.parse(
-        '$kZhjwBase/student/integratedQuery/scoreQuery/allPassingScores/index',
-      ),
-      headers: {
-        'Accept':
-            'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-        'Referer': '$kZhjwBase/',
-        'User-Agent': _headers['User-Agent']!,
-      },
-    );
-    final indexBody = indexResp.body;
-    _checkSessionExpiry(indexBody, indexResp.statusCode);
-    final urlMatch = RegExp(
-      r'var\s+url\s*=\s*"(/student/integratedQuery/scoreQuery/[^/]+/allPassingScores/callback)"',
-    ).firstMatch(indexBody);
-    if (urlMatch == null) {
-      if (indexBody.contains('login') || indexBody.contains('Login')) {
-        throw ScuLoginException('登录已过期，请重新登录', sessionExpired: true);
-      }
-      throw ScuLoginException('无法从页面提取 allPassingScores callback URL');
-    }
-    final callbackPath = urlMatch.group(1)!;
-
-    final callbackResp = await client.get(
-      Uri.parse('$kZhjwBase$callbackPath'),
-      headers: {
-        'Accept': 'application/json, text/plain, */*',
-        'Referer':
-            '$kZhjwBase/student/integratedQuery/scoreQuery/allPassingScores/index',
-        'User-Agent': _headers['User-Agent']!,
-      },
-    );
-    final body = callbackResp.body.trim();
-    _checkSessionExpiry(body, callbackResp.statusCode);
-    return parseJson(
-      body,
-      'allPassingScores/callback',
-      (msg) => ScuLoginException(msg),
-    );
-  }
-
-  /// 获取方案成绩（需要先登录）
-  Future<Map<String, dynamic>> fetchSchemeScores({CookieClient? client}) async {
-    client ??= await bindSession();
-    final indexResp = await client.get(
-      Uri.parse(
-        '$kZhjwBase/student/integratedQuery/scoreQuery/schemeScores/index',
-      ),
-      headers: {
-        'Accept':
-            'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-        'Referer': '$kZhjwBase/',
-        'User-Agent': _headers['User-Agent']!,
-      },
-    );
-
-    final indexBody = indexResp.body;
-    _checkSessionExpiry(indexBody, indexResp.statusCode);
-
-    final urlMatch = RegExp(
-      r'var\s+url\s*=\s*"(/student/integratedQuery/scoreQuery/[^/]+/schemeScores/callback)"',
-    ).firstMatch(indexBody);
-    if (urlMatch == null) {
-      if (indexBody.contains('login') || indexBody.contains('Login')) {
-        throw ScuLoginException('登录已过期，请重新登录', sessionExpired: true);
-      }
-      throw ScuLoginException('无法从页面提取 schemeScores callback URL');
-    }
-    final callbackPath = urlMatch.group(1)!;
-    final callbackResp = await client.get(
-      Uri.parse('$kZhjwBase$callbackPath'),
-      headers: {
-        'Accept': 'application/json, text/plain, */*',
-        'Referer':
-            '$kZhjwBase/student/integratedQuery/scoreQuery/schemeScores/index',
-        'User-Agent': _headers['User-Agent']!,
-      },
-    );
-
-    final body = callbackResp.body.trim();
-    _checkSessionExpiry(body, callbackResp.statusCode);
-    return parseJson(
-      body,
-      'schemeScores/callback',
-      (msg) => ScuLoginException(msg),
-    );
-  }
-
-  /// 获取教室查询页面的校区和教学楼列表
-  Future<({List<ClassroomCampus> campuses, List<ClassroomBuilding> buildings})>
-  fetchClassroomIndex({CookieClient? client}) async {
-    client ??= await bindSession();
-    final resp = await client.get(
-      Uri.parse(
-        '$kZhjwBase/student/teachingResources/classroomUseStatus/index',
-      ),
-      headers: {
-        'Accept': 'text/html,*/*',
-        'Referer': '$kZhjwBase/',
-        'User-Agent': _headers['User-Agent']!,
-      },
-    );
-    final body = resp.body.trim();
-    _checkSessionExpiry(body, resp.statusCode);
-    final xqMatch = RegExp(
-      r"""<input[^>]+id="xqList"[^>]+value='([^']+)'""",
-    ).firstMatch(body);
-    if (xqMatch == null) {
-      throw ScuLoginException('无法解析校区列表');
-    }
-    final xqList = (jsonDecode(xqMatch.group(1)!) as List)
-        .map((e) => ClassroomCampus.fromJson(e as Map<String, dynamic>))
-        .toList();
-
-    final jxlMatch = RegExp(
-      r"""<input[^>]+id="jxlList"[^>]+value='([^']+)'""",
-    ).firstMatch(body);
-    if (jxlMatch == null) {
-      throw ScuLoginException('无法解析教学楼列表');
-    }
-    final jxlList = (jsonDecode(jxlMatch.group(1)!) as List)
-        .map((e) => ClassroomBuilding.fromJson(e as Map<String, dynamic>))
-        .toList();
-
-    return (campuses: xqList, buildings: jxlList);
-  }
-
-  /// 获取教学楼的教室类型列表
-  Future<List<ClassroomType>> fetchClassroomTypes({
-    required String campusNumber,
-    required String buildingNumber,
-    required String campusName,
-    required String buildingName,
-    CookieClient? client,
-  }) async {
-    client ??= await bindSession();
-    final resp = await client.get(
-      Uri.parse(
-        '$kZhjwBase/student/teachingResources/classroomUseStatus'
-        '/$campusNumber/$buildingNumber'
-        '/${Uri.encodeComponent(campusName)}/${Uri.encodeComponent(buildingName)}',
-      ),
-      headers: {
-        'Accept': 'text/html,*/*',
-        'Referer': '$kZhjwBase/',
-        'User-Agent': _headers['User-Agent']!,
-      },
-    );
-    final body = resp.body.trim();
-    _checkSessionExpiry(body, resp.statusCode);
-    final match = RegExp(
-      r"""<input[^>]+id="classroomTypes"[^>]+value='([^']+)'""",
-    ).firstMatch(body);
-    if (match == null) return [];
-    return (jsonDecode(match.group(1)!) as List)
-        .map((e) => ClassroomType.fromJson(e as Map<String, dynamic>))
-        .toList();
-  }
-
-  /// 查询教室使用情况
-  Future<ClassroomQueryResult> fetchClassroomAvailability({
-    required String campusNumber,
-    required String buildingNumber,
-    String classroomType = '',
-    String classroomName = '',
-    String seatFrom = '',
-    String seatTo = '',
-    String searchDate = '',
-    CookieClient? client,
-  }) async {
-    client ??= await bindSession();
-    final resp = await client.post(
-      Uri.parse(
-        '$kZhjwBase/student/teachingResources/classroomUseStatus/jasInfo',
-      ),
-      headers: {
-        'Accept': 'application/json, text/javascript, */*; q=0.01',
-        'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8',
-        'Referer':
-            '$kZhjwBase/student/teachingResources/classroomUseStatus/index',
-        'User-Agent': _headers['User-Agent']!,
-        'X-Requested-With': 'XMLHttpRequest',
-      },
-      body:
-          'xqh=${Uri.encodeComponent(campusNumber)}'
-          '&jxlh=${Uri.encodeComponent(buildingNumber)}'
-          '&jslx=${Uri.encodeComponent(classroomType)}'
-          '&jasm=${Uri.encodeComponent(classroomName)}'
-          '&zwFrom=${Uri.encodeComponent(seatFrom)}'
-          '&zwTo=${Uri.encodeComponent(seatTo)}'
-          '&searchDate=${Uri.encodeComponent(searchDate)}',
-    );
-    final body = resp.body.trim();
-    _checkSessionExpiry(body, resp.statusCode);
-    return ClassroomQueryResult.fromJson(
-      parseJson(
-        body,
-        'classroomUseStatus/jasInfo',
-        (msg) => ScuLoginException(msg),
-      ),
-    );
-  }
-
-  void logout() {
-    _accessToken = null;
-    _cachedClient = null;
-  }
-
-  /// 检查会话是否过期，过期则抛出 [ScuLoginException]。
-  void _checkSessionExpiry(String body, int statusCode) {
-    if (statusCode == 302) {
-      throw ScuLoginException('登录已过期，请重新登录', sessionExpired: true);
-    }
-    if (body.trim().isEmpty) {
-      throw ScuLoginException('登录已过期，请重新登录', sessionExpired: true);
-    }
-    if (body.startsWith('<') && body.contains('login')) {
-      throw ScuLoginException('登录已过期，请重新登录', sessionExpired: true);
-    }
-  }
-}
-
-class CaptchaResult {
-  final String code;
-  final String captchaBase64;
-  const CaptchaResult({required this.code, required this.captchaBase64});
-}
-
-class ScuLoginException implements Exception {
-  final String message;
-  final bool sessionExpired;
-  const ScuLoginException(this.message, {this.sessionExpired = false});
-  @override
-  String toString() => message;
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Cookie 感知的 http.Client
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Cookie 感知的 http.Client，按域名隔离存储，发送时只带当前请求域的 cookie。
-class CookieClient extends http.BaseClient {
-  // 在 Ubuntu 24.04 Desktop 中曾出现：教务系统对 Keep-Alive 连接比较脆弱，不能稳定复用连接的现象，
-  // 导致从教务系统获取信息失败，表现为程序输出报错日志：
-  //   Scheme scores load error: ClientException: 断开的管道, uri=http://zhjw.scu.edu.cn/index
-  // 这里修复为：增加重试机制，如果教务系统管道断开就重新连接并发送请求，
-  // 在 _inner.send 外再套一层重试的壳：sendWithClientExceptionRetry
-  http.Client _inner = http.Client();
-
-  // 按域名存 cookie：host -> {name: value}
-  final _jar = <String, Map<String, String>>{};
-
-  bool _reusable = false;
-
-  /// 返回所有域 cookie 拼接的字符串（仅用于调试日志）
-  String get cookieHeader {
-    final all = <String, String>{};
-    for (final m in _jar.values) {
-      all.addAll(m);
-    }
-    return all.entries.map((e) => '${e.key}=${e.value}').join('; ');
-  }
-
-  /// 获取适合发送给 [uri] 的 cookie（域名匹配：精确 host 或父域）
-  Map<String, String> _cookiesFor(Uri uri) {
-    final host = uri.host; // e.g. "id.scu.edu.cn"
-    final result = <String, String>{};
-    for (final entry in _jar.entries) {
-      final jarHost = entry.key;
-      // 精确匹配 或 host 是 jarHost 的子域（.scu.edu.cn 匹配 id.scu.edu.cn）
-      if (host == jarHost || host.endsWith('.$jarHost')) {
-        result.addAll(entry.value);
-      }
-    }
-    return result;
-  }
-
-  /// 解析并存储响应中的 Set-Cookie
-  void _storeCookies(Uri uri, http.BaseResponse response) {
-    final raw = response.headers['set-cookie'];
-    if (raw == null) return;
-
-    final host = uri.host;
-    _jar.putIfAbsent(host, () => {});
-
-    // 按 ", " 分割，但跳过 expires 日期里的逗号（如 "Thu, 01 Jan 2099"）
-    for (final part in raw.split(RegExp(r',\s*(?=[A-Za-z][^,=\s]*\s*=)'))) {
-      final kv = part.split(';').first.trim();
-      final eq = kv.indexOf('=');
-      if (eq > 0) {
-        final name = kv.substring(0, eq).trim();
-        final value = kv.substring(eq + 1).trim();
-        _jar[host]![name] = value;
-      }
-    }
-  }
-
-  /// 手动跟随重定向，每跳都：
-  ///   1. 只带当前跳目标域的 cookie
-  ///   2. 收集响应的 Set-Cookie 存回对应域
-  Future<http.Response> followRedirects(
-    Uri url, {
-    Map<String, String>? headers,
-    int maxRedirects = 10,
-  }) async {
-    Uri current = url;
-    http.Response? lastResponse;
-
-    for (int i = 0; i <= maxRedirects; i++) {
-      // 只发当前域的 cookie，而非全部域
-      final cookies = _cookiesFor(current);
-      final reqHeaders = <String, String>{
-        ...?headers,
-        if (cookies.isNotEmpty)
-          'Cookie': cookies.entries
-              .map((e) => '${e.key}=${e.value}')
-              .join('; '),
-      };
-
-      final request = http.Request('GET', current)
-        ..followRedirects = false
-        ..headers.addAll(reqHeaders);
-
-      // 如果教务系统管道断开，就重新连接并发送请求
-      final streamed = await sendWithClientExceptionRetry(request);
-      final response = await http.Response.fromStream(streamed);
-      _storeCookies(current, response);
-
-      if (response.statusCode >= 300 && response.statusCode < 400) {
-        final location = response.headers['location'];
-        if (location == null) break;
-        current = current.resolve(location);
-        lastResponse = response;
-      } else {
-        return response;
-      }
-    }
-    return lastResponse!;
-  }
-
-  @override
-  Future<http.StreamedResponse> send(http.BaseRequest request) async {
-    final cookies = _cookiesFor(request.url);
-    if (cookies.isNotEmpty) {
-      request.headers['Cookie'] = cookies.entries
-          .map((e) => '${e.key}=${e.value}')
-          .join('; ');
-    }
-    final response = await _inner.send(request).timeout(kHttpTimeout);
-    _storeCookies(request.url, response);
-    return response;
-  }
-
-  /// 带 http.ClientException 错误重试的请求发送函数
-  Future<http.StreamedResponse> sendWithClientExceptionRetry(
-    http.BaseRequest request,
-  ) async {
-    try {
-      return await _inner.send(request).timeout(kHttpTimeout);
-    } on http.ClientException catch (_) {
-      _inner.close();
-      _inner = http.Client();
-      final retryRequest = http.Request(request.method, request.url)
-        ..followRedirects = request.followRedirects
-        ..maxRedirects = request.maxRedirects
-        ..persistentConnection = true
-        ..headers.addAll(request.headers);
-      if (request is http.Request) {
-        retryRequest.body = request.body;
-      }
-      return await _inner.send(retryRequest).timeout(kHttpTimeout);
-    }
-  }
-
-  @override
-  void close() {
-    if (_reusable) return;
-    _inner.close();
-    super.close();
-  }
-}
+// Barrel export — 保持旧 import 路径兼容。
+export 'package:bugaoshan/services/scu_auth/scu_auth_service.dart';
+export 'package:bugaoshan/services/scu_auth/scu_auth_models.dart';
+export 'package:bugaoshan/services/scu_auth/cookie_client.dart';
